+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.2.0
+
+Moved fonts to `fonts/` folder; this directory should now be included with the package.
+Work to support referencing these fonts when `figlet` is used as dependency
+Included fonts are now identified by their relative path, e.g. "figlet.js/Alpha.flf"
+Various documentation improvements
+
 ## v0.1.0
 
 Initial POC release. Lacks support for screen widths, unicode characters, and text squashing.